+++ rvp64_exec/builtin-rvp64-smmul.c
@@ -1,15 +1,15 @@
 /* This is a test program for smmul instruction.  */
 /* { dg-do run } */
-/* { dg-options "-march=rv64gc_zpn_zprv_zpsf -mabi=lp64d -O2" } */
+/* { dg-options "-march=rv64gc_zpn_zpsf -mabi=lp64d -O2" } */
 
 
 #include <rvp_intrinsic.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "test_utils.h"
+
 
 static __attribute__ ((noinline))
-int32_t smmul (int32_t ra, int32_t rb)
+int64_t smmul (int64_t ra, int64_t rb)
 {
   return __rv_smmul (ra, rb);
 }
@@ -17,7 +17,7 @@ int32_t smmul (int32_t ra, int32_t rb)
 int
 main ()
 {
-  int32_t a = smmul (0x8000000080000000, 0x8000000080000000);
+  int64_t a = smmul (0x8000000080000000, 0x8000000080000000);
 
   if (a != 0x4000000040000000)
     abort ();
