+++ rvp64_exec/builtin-rvp64-smmwtu.c
@@ -1,15 +1,15 @@
 /* This is a test program for smmwtu instruction.  */
 /* { dg-do run } */
-/* { dg-options "-march=rv64gc_zpn_zprv_zpsf -mabi=lp64d -O2" } */
+/* { dg-options "-march=rv64gc_zpn_zpsf -mabi=lp64d -O2" } */
 
 
 #include <rvp_intrinsic.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "test_utils.h"
+
 
 static __attribute__ ((noinline))
-int32_t smmwtu (int32_t ra, int32_t rb)
+int64_t smmwtu (int64_t ra, int64_t rb)
 {
   return __rv_smmwt_u (ra, rb);
 }
@@ -23,7 +23,7 @@ int32x2_t v_smmwtu (int32x2_t ra, int16x4_t rb)
 int
 main ()
 {
-  int32_t a = smmwtu (0x8000000080000000, 0x8000000180000001);
+  int64_t a = smmwtu (0x8000000080000000, 0x8000000180000001);
   int32x2_t va = v_smmwtu ((int32x2_t) {0x80000000, 0x80000000},
 			   (int16x4_t) {0x1, 0x8000, 0x1, 0x8000});
